+++ packages/capture-kit/src/react-native-overlay.ts
