+++ packages/capture-kit/src/screenshot-overlay-rects.ts
